Move checkMissingPod func out of controller pkg.
This commit moves the checkMissingPod function from controller
to the status package and become a exported function. This
change allows CheckMissingPod function to stay closer with
relevant content and could be used across the packages.
