+++ controllers/loadtest_controller_test.go
@@ -9,7 +9,6 @@ import (
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 var _ = Describe("Test Environment", func() {
@@ -697,147 +696,6 @@ var _ = Describe("Pod Creation", func() {
 		It("grants read and write access", func() {
 			volumeMount := newWorkspaceVolumeMount()
 			Expect(volumeMount.ReadOnly).To(BeFalse())
-		})
-	})
-})
-
-var _ = Describe("checkMissingPods", func() {
-
-	var currentLoadTest *grpcv1.LoadTest
-	var allRunningPods *corev1.PodList
-	var actualReturn *LoadTestMissing
-	var expectedReturn *LoadTestMissing
-
-	BeforeEach(func() {
-		currentLoadTest = newLoadTestWithMultipleClientsAndServers()
-		newListRef := corev1.PodList{Items: []corev1.Pod{}}
-		allRunningPods = &newListRef
-		expectedReturn = &LoadTestMissing{Clients: []grpcv1.Client{}, Servers: []grpcv1.Server{}}
-	})
-
-	Describe("no pods from the current load test is running", func() {
-		BeforeEach(func() {
-			for i := 0; i < len(currentLoadTest.Spec.Clients); i++ {
-				expectedReturn.Clients = append(expectedReturn.Clients, currentLoadTest.Spec.Clients[i])
-			}
-			for i := 0; i < len(currentLoadTest.Spec.Servers); i++ {
-				expectedReturn.Servers = append(expectedReturn.Servers, currentLoadTest.Spec.Servers[i])
-			}
-			expectedReturn.Driver = currentLoadTest.Spec.Driver
-		})
-
-		Context("no pod is running", func() {
-			It("returns the full pod list from the current load test", func() {
-				actualReturn = checkMissingPods(currentLoadTest, allRunningPods)
-				Expect(actualReturn.Clients).To(ConsistOf(expectedReturn.Clients))
-				Expect(actualReturn.Servers).To(ConsistOf(expectedReturn.Servers))
-				Expect(actualReturn.Driver).To(Equal(expectedReturn.Driver))
-			})
-		})
-
-		Context("irrelevant pods are running", func() {
-			It("returns the full pod list from the current load test", func() {
-				allRunningPods.Items = append(allRunningPods.Items, createPodListWithIrrelevantPod().Items...)
-				actualReturn = checkMissingPods(currentLoadTest, allRunningPods)
-				Expect(actualReturn.Clients).To(ConsistOf(expectedReturn.Clients))
-				Expect(actualReturn.Servers).To(ConsistOf(expectedReturn.Servers))
-				Expect(actualReturn.Driver).To(Equal(expectedReturn.Driver))
-			})
-		})
-	})
-
-	Describe("some of pods from the current load test is running", func() {
-
-		BeforeEach(func() {
-			allRunningPods.Items = append(allRunningPods.Items,
-				corev1.Pod{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "random-name",
-						Labels: map[string]string{
-							config.LoadTestLabel:      "test-loadtest-multiple-clients-and-servers",
-							config.RoleLabel:          "server",
-							config.ComponentNameLabel: "server-1",
-						},
-					},
-				},
-				corev1.Pod{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "random-name",
-						Labels: map[string]string{
-							config.LoadTestLabel:      "test-loadtest-multiple-clients-and-servers",
-							config.RoleLabel:          "client",
-							config.ComponentNameLabel: "client-2",
-						},
-					},
-				},
-				corev1.Pod{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: "random-name",
-						Labels: map[string]string{
-							config.LoadTestLabel:      "test-loadtest-multiple-clients-and-servers",
-							config.RoleLabel:          "driver",
-							config.ComponentNameLabel: "driver-1",
-						},
-					},
-				},
-			)
-			for i := 0; i < len(currentLoadTest.Spec.Clients); i++ {
-				if *currentLoadTest.Spec.Clients[i].Name != "client-2" {
-					expectedReturn.Clients = append(expectedReturn.Clients, currentLoadTest.Spec.Clients[i])
-				}
-			}
-
-			for i := 0; i < len(currentLoadTest.Spec.Servers); i++ {
-				if *currentLoadTest.Spec.Servers[i].Name != "server-1" {
-					expectedReturn.Servers = append(expectedReturn.Servers, currentLoadTest.Spec.Servers[i])
-				}
-			}
-		})
-
-		Context("only pods from the current load test are running", func() {
-			It("returns the list of pods missing from collection of running pods", func() {
-				actualReturn = checkMissingPods(currentLoadTest, allRunningPods)
-				Expect(actualReturn.Clients).To(ConsistOf(expectedReturn.Clients))
-				Expect(actualReturn.Servers).To(ConsistOf(expectedReturn.Servers))
-				Expect(actualReturn.Driver).To(Equal(expectedReturn.Driver))
-			})
-		})
-
-		Context("there are also irrelevant pods running", func() {
-			It("returns the list of pods missing from collection of running pods", func() {
-				allRunningPods.Items = append(allRunningPods.Items, createPodListWithIrrelevantPod().Items...)
-				actualReturn = checkMissingPods(currentLoadTest, allRunningPods)
-				Expect(actualReturn.Clients).To(ConsistOf(expectedReturn.Clients))
-				Expect(actualReturn.Servers).To(ConsistOf(expectedReturn.Servers))
-				Expect(actualReturn.Driver).To(Equal(expectedReturn.Driver))
-			})
-		})
-	})
-
-	Describe("all of pods from the current load test is running", func() {
-
-		BeforeEach(func() {
-			allRunningPods = populatePodListWithCurrentLoadTestPod(currentLoadTest)
-		})
-
-		Context("only pods from the current load test are running", func() {
-			It("returns a empty list", func() {
-				actualReturn = checkMissingPods(currentLoadTest, allRunningPods)
-				Expect(actualReturn.Clients).To(ConsistOf(expectedReturn.Clients))
-				Expect(actualReturn.Servers).To(ConsistOf(expectedReturn.Servers))
-				Expect(actualReturn.Driver).To(Equal(expectedReturn.Driver))
-			})
-		})
-
-		Context("there are also irrelevant pods running", func() {
-
-			It("returns an empty list", func() {
-				allRunningPods.Items = append(allRunningPods.Items, createPodListWithIrrelevantPod().Items...)
-				actualReturn = checkMissingPods(currentLoadTest, allRunningPods)
-				Expect(actualReturn.Clients).To(ConsistOf(expectedReturn.Clients))
-				Expect(actualReturn.Servers).To(ConsistOf(expectedReturn.Servers))
-				Expect(actualReturn.Driver).To(Equal(expectedReturn.Driver))
-			})
 		})
 	})
 })
