+++ status/suite_test.go
@@ -19,11 +19,170 @@ package status
 import (
 	"testing"
 
+	grpcv1 "github.com/grpc/test-infra/api/v1"
+	"github.com/grpc/test-infra/config"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
+	corev1 "k8s.io/api/core/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 )
 
 func TestStatus(t *testing.T) {
 	RegisterFailHandler(Fail)
 	RunSpecs(t, "Status Suite")
+}
+
+// newLoadTestWithMultipleClientsAndServers attempts to create a
+// loadtest with multiple clients and servers for test.
+func newLoadTestWithMultipleClientsAndServers() *grpcv1.LoadTest {
+	cloneImage := "docker.pkg.github.com/grpc/test-infra/clone"
+	cloneRepo := "https://github.com/grpc/grpc.git"
+	cloneGitRef := "master"
+
+	buildImage := "l.gcr.io/google/bazel:latest"
+	buildCommand := []string{"bazel"}
+	buildArgs := []string{"build", "//test/cpp/qps:qps_worker"}
+
+	driverImage := "docker.pkg.github.com/grpc/test-infra/driver"
+	runImage := "docker.pkg.github.com/grpc/test-infra/cxx"
+	runCommand := []string{"bazel-bin/test/cpp/qps/qps_worker"}
+
+	clientRunArgs := []string{"--driver_port=10000"}
+	serverRunArgs := append(clientRunArgs, "--server_port=10010")
+
+	bigQueryTable := "grpc-testing.e2e_benchmark.foobarbuzz"
+
+	driverPool := "drivers"
+	workerPool := "workers-8core"
+
+	driverComponentName := "driver-1"
+
+	createdLoadTest := &grpcv1.LoadTest{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      "test-loadtest-multiple-clients-and-servers",
+			Namespace: "default",
+		},
+
+		Spec: grpcv1.LoadTestSpec{
+			Driver: &grpcv1.Driver{
+				Component: grpcv1.Component{
+					Name:     &driverComponentName,
+					Language: "cxx",
+					Pool:     &driverPool,
+					Run: grpcv1.Run{
+						Image: &driverImage,
+					},
+				},
+			},
+			Results: &grpcv1.Results{
+				BigQueryTable: &bigQueryTable,
+			},
+
+			Scenarios: []grpcv1.Scenario{
+				{Name: "cpp-example-scenario"},
+			},
+		},
+	}
+
+	serverNames := []string{"server-1", "server-2", "server-3"}
+	for i := 1; i <= len(serverNames); i++ {
+		createdLoadTest.Spec.Servers = append(createdLoadTest.Spec.Servers, grpcv1.Server{
+			Component: grpcv1.Component{
+				Name:     &serverNames[i-1],
+				Language: "cxx",
+				Pool:     &workerPool,
+				Clone: &grpcv1.Clone{
+					Image:  &cloneImage,
+					Repo:   &cloneRepo,
+					GitRef: &cloneGitRef,
+				},
+				Build: &grpcv1.Build{
+					Image:   &buildImage,
+					Command: buildCommand,
+					Args:    buildArgs,
+				},
+				Run: grpcv1.Run{
+					Image:   &runImage,
+					Command: runCommand,
+					Args:    serverRunArgs,
+				},
+			},
+		})
+	}
+
+	clientName := []string{"client-1", "client-2", "client-3"}
+	for i := 1; i <= len(clientName); i++ {
+		createdLoadTest.Spec.Clients = append(createdLoadTest.Spec.Clients, grpcv1.Client{
+			Component: grpcv1.Component{
+				Name:     &clientName[i-1],
+				Language: "cxx",
+				Pool:     &workerPool,
+				Clone: &grpcv1.Clone{
+					Image:  &cloneImage,
+					Repo:   &cloneRepo,
+					GitRef: &cloneGitRef,
+				},
+				Build: &grpcv1.Build{
+					Image:   &buildImage,
+					Command: buildCommand,
+					Args:    buildArgs,
+				},
+				Run: grpcv1.Run{
+					Image:   &runImage,
+					Command: runCommand,
+					Args:    clientRunArgs,
+				},
+			},
+		})
+	}
+
+	return createdLoadTest
+}
+
+// populatePodListWithCurrentLoadTestPod attempts to create a Podlist and populate
+// it with the pods came from current loadtest.
+func populatePodListWithCurrentLoadTestPod(currentLoadTest *grpcv1.LoadTest) []*corev1.Pod {
+	var currentPodList []*corev1.Pod
+
+	for _, eachClient := range currentLoadTest.Spec.Clients {
+		currentPodList = append(currentPodList,
+			&corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "random-name",
+					Labels: map[string]string{
+						config.LoadTestLabel:      currentLoadTest.Name,
+						config.RoleLabel:          "client",
+						config.ComponentNameLabel: *eachClient.Name,
+					},
+				},
+			})
+	}
+
+	for _, eachServer := range currentLoadTest.Spec.Servers {
+		currentPodList = append(currentPodList,
+			&corev1.Pod{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: "random-name",
+					Labels: map[string]string{
+						config.LoadTestLabel:      currentLoadTest.Name,
+						config.RoleLabel:          "server",
+						config.ComponentNameLabel: *eachServer.Name,
+					},
+				},
+			})
+	}
+
+	currentPodList = append(currentPodList,
+		&corev1.Pod{
+			ObjectMeta: metav1.ObjectMeta{
+				Name: "random-name",
+				Labels: map[string]string{
+					config.LoadTestLabel:      currentLoadTest.Name,
+					config.RoleLabel:          "driver",
+					config.ComponentNameLabel: *currentLoadTest.Spec.Driver.Name,
+				},
+			},
+		})
+
+	return currentPodList
 }
